feat(商家,购物车,订单):爬取数据SQL,完成购物车,完成订单/下单

+++ src/main/java/com/lsd/fun/modules/cms/entity/CouponOrderEntity.java
@@ -20,7 +20,7 @@ public class CouponOrderEntity implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	/**
-	 * 
+	 *
 	 */
 	@TableId
 	private Integer id;
@@ -35,7 +35,7 @@ public class CouponOrderEntity implements Serializable {
 	/**
 	 * 商品订单ID
 	 */
-	private Integer orderId;
+	private Long orderId;
 	/**
 	 * 创建时间
 	 */
